Add `HostnamePolicy` to `Controlplane` configuration
Extend the `Controlplane` provider configuration to allow to set the
`HostnamePolicy` for a `Shoot` cluster. Supported values are `Server`.
If that field is set, the MCM gets an extra `arg` to use the `Server`
name as a `Node` name instead of using the `ServerClaim` name.

If this field is not set, `ServerClaim` name will be used (default
behaviour).